requirements.txt
# ===== CODE GENERATION =====
python-dotenv>=1.0.0,<2.0.0
datamodel-code-generator>=0.16.0,<1.0.0
# ===== SynapScale Backend - Requirements Completo =====
# Arquivo unificado de dependências - 12/06/2025
# Combinado a partir de requirements.txt, requirements.backend.txt e requirements.notorch.txt

# ===== CORE FRAMEWORK =====
fastapi>=0.110.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.30.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
gunicorn>=21.2.0,<22.0.0

# ===== BANCO DE DADOS =====
sqlalchemy>=2.0.25,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0

# ===== CACHE E REDIS =====
redis>=5.0.1,<6.0.0
aioredis>=2.0.1,<3.0.0

# ===== AUTENTICAÇÃO E SEGURANÇA =====
pyjwt>=2.8.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib>=1.7.4
python-multipart>=0.0.6,<0.1.0
cryptography>=42.0.0,<43.0.0
bcrypt>=4.1.2

# ===== HTTP E REQUESTS =====
httpx>=0.26.0,<0.28.0
aiohttp>=3.9.3,<4.0.0
requests>=2.31.0,<3.0.0

# ===== PROCESSAMENTO DE DADOS =====
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<2.0.0
python-dateutil>=2.8.2,<3.0.0

# ===== VALIDAÇÃO E SERIALIZAÇÃO =====
email-validator>=2.1.0,<3.0.0
python-slugify>=8.0.2,<9.0.0
pydantic-extra-types>=2.5.0,<3.0.0

# ===== LOGGING E MONITORAMENTO =====
loguru>=0.7.2,<0.8.0
sentry-sdk[fastapi]>=1.40.0,<2.0.0
prometheus-client>=0.19.0,<0.21.0
prometheus-fastapi-instrumentator>=6.1.0,<7.0.0
psutil>=5.9.0,<6.0.0

# ===== UTILITÁRIOS =====
python-dotenv>=1.0.1,<2.0.0
click>=8.1.7,<9.0.0
rich>=13.7.0,<14.0.0
typer>=0.9.0,<0.13.0
jsonpath-ng>=1.6.0,<2.0.0

# ===== RATE LIMITING =====
slowapi>=0.1.9,<0.2.0

# ===== WEBSOCKETS =====
websockets>=12.0,<13.0

# ===== UPLOAD E ARQUIVOS =====
python-magic>=0.4.27,<0.5.0
pillow>=10.2.0,<11.0.0
aiofiles>=23.2.1,<24.0.0

# ===== TEMPLATES E EMAIL =====
jinja2>=3.1.3,<4.0.0
aiosmtplib>=3.0.1,<4.0.0

# ===== TIMEZONE =====
pytz>=2024.1

# ===== TESTING (DESENVOLVIMENTO) =====
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.5,<0.24.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.12.0,<4.0.0
factory-boy>=3.3.0,<4.0.0

# ===== CODE QUALITY =====
black>=24.2.0,<25.0.0
isort>=5.13.2,<6.0.0
mypy>=1.8.0,<2.0.0
flake8>=7.0.0,<8.0.0
pre-commit>=3.6.0,<4.0.0

# ===== PROVEDORES DE IA =====
# OpenAI
openai>=1.12.0,<2.0.0
tiktoken>=0.5.2,<1.0.0

# Anthropic Claude
anthropic>=0.15.0,<1.0.0

# Google Gemini
google-generativeai>=0.4.0,<1.0.0

# Groq
groq>=0.4.2,<1.0.0

# Cohere
cohere>=4.47.0,<5.0.0

# DeepSeek (em desenvolvimento)
#deepseek-ai>=1.0.0,<2.0.0

# Together AI
together>=0.2.7,<1.0.0

# Replicate
replicate>=0.25.0,<1.0.0

# ===== HUGGINGFACE E ML =====
transformers>=4.38.0,<5.0.0
accelerate>=0.27.0,<1.0.0

# PyTorch (descomente conforme necessário)
# torch>=2.2.0,<3.0.0  # Descomente se precisar de PyTorch

# ===== VECTOR DATABASES =====
pinecone-client>=3.0.0,<4.0.0
chromadb>=0.4.22,<1.0.0
weaviate-client>=4.4.0,<5.0.0

# ===== DOCUMENT PROCESSING =====
pypdf>=4.0.0,<5.0.0
python-docx>=1.1.0,<2.0.0
openpyxl>=3.1.2,<4.0.0
markdown>=3.5.2,<4.0.0

# ===== CLOUD STORAGE =====
boto3>=1.34.0,<2.0.0
google-cloud-storage>=2.14.0,<3.0.0
azure-storage-blob>=12.19.0,<13.0.0

# ===== PRODUÇÃO E DEPLOYMENT =====
# Descomente conforme necessário para ambientes de produção
# newrelic>=9.5.0,<10.0.0
# datadog>=0.49.1,<1.0.0
# elastic-apm>=6.20.0,<7.0.0
